Refactor DelayedResponse and derived classes
Move logic from DelayedResponse to derived classes to make it a simple,
universal interface. This gives derived classes much more flexibility
about how the actual logic is implemented and makes the interface easier
to understand.

In addition, use QPointer to be able detecting if one of the involved
widget gets deleted during the execution of the delayed events.